[Radoub] refactor: Promote WindowLifecycleManager to Radoub.UI (#2391)
Move the generic singleton-window manager from Parley (DialogEditor.Services)
to Radoub.UI.Services so other tools can share it. Parley-specific WindowKeys
stays in the tool. Manager tests move to Radoub.UI.Tests; WindowKeys test stays
in Parley.Tests. Refactor-only, no behavior change.
Tests: Radoub.UI.Tests 14/14, Parley.Tests WindowKeys 1/1.
